Min ABS Difference
Consider an array of integers, . We define the absolute difference
between two elements, and  (where ), to be the absolute value of .
Given an array of  integers, find and print the minimum absolute
difference between any two elements in the array.
Input Format
The first line contains a single integer denoting  (the number of
integers).
The second line contains  space-separated integers describing the
respective values of .
Constraints
"
"
Output Format
Print the minimum absolute difference between any two elements in the
array.
Sample Input 0
3
3 -7 0
Sample Output 0
3
Explanation 0
With n=3 integers in our array, we have three possible pairs: (3,-7)
,(3,0) and (-7,0). The absolute values of the differences between these
pairs are as follows:
|3 - - 7| = 10
|3 - 0| = 3
|-7 - 0| = 7
Notice that if we were to switch the order of the numbers in these
pairs, the resulting absolute values would still be the same. The
smallest of these possible absolute differences is 3, so we print 3 as
our answer