Funny String
In this challenge, you will determine whether a string is funny or not.
To determine whether a string is funny, create a copy of the string in
reverse e.g. . Iterating through each string, compare the absolute
difference in the ascii values of the characters at positions 0 and 1,
1 and 2 and so on to the end. If the list of absolute differences is
the same for both strings, they are funny.

Note: The absolute value of , is written as .

Input Format

The first line contains an integer , (the number of test cases).
The next  lines contain a string, .

Constraints

Output Format

For each string  print whether it is  or  on a new line.

Sample Input

2
acxz
bcxz
Sample Output

Funny
Not Funny
Explanation

You can use  to store the reverse of .

Test Case 0:

Corresponding ASCII values of characters of the strings:
 and
For both the strings the adjacent difference list is [2, 21, 2] so we
print .

Test Case 1:

Corresponding ASCII values of characters of the strings:
 and
The adjacent difference list for string s is [1, 21, 2] and for string
r it is [2, 21, 1]. Since they are not the same we print Not Funny